requirements.txt
runpod>=1.7.0
pydantic>=2.7.0,<3
PyYAML>=6.0
